+++ pytket/pytket/passes/auto_rebase.py
@@ -134,6 +134,10 @@ def auto_rebase_pass(gateset: Set[OpType], allow_swaps: bool = False) -> BasePas
     Raises an error if no known decompositions can be found, in which case try
     using RebaseCustom with your own decompositions.
 
+    In addition to the gate types in ``gateset``, any ``Measure``, ``Reset`` and
+    ``Collapse`` operations in the original circuit are retained. Conditional
+    operations are also allowed. ``Phase`` gates may also be introduced.
+
     :param gateset: Set of supported OpTypes, target gate set.
     :type gateset: FrozenSet[OpType]
     :raises NoAutoRebase: No suitable decomposition found.