+++ pytket/docs/changelog.rst
@@ -1,6 +1,29 @@
 Changelog
 =========
 
+1.23.0 (January 2024)
+---------------------
+
+API changes:
+
+* Make the ``architecture`` field in ``BackendInfo`` optional.
+
+Deprecations:
+
+* Deprecate ``SynthesiseHQS`` pass.
+  
+Fixes:
+
+* Ensure that squashing long sequences of gates via unitary multiplication does
+  not produce non-unitary results due to rounding errors.
+* Fix `PauliFrameRandomisation.sample_circuits`.
+* For `Circuit` with no 2-qubit gates, `NoiseAwarePlacement` now assigns `Qubit` to `Node` in `Architecture`
+  with lowest reported error rates.
+* Fix invalid registers returned by ``Circuit.q_registers`` and ``Circuit.c_registers``.
+* Fix regression (introduced in 1.22.0) in compilation performance with certain
+  sequences of passes.
+
+
 1.22.0 (November 2023)
 ----------------------